Revert change to MAX_HISTORY for isUnlimited().

+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/limit/FaweLimit.java
@@ -235,7 +235,7 @@ public class FaweLimit {
                 && MAX_ITERATIONS == Integer.MAX_VALUE
                 && MAX_BLOCKSTATES == Integer.MAX_VALUE
                 && MAX_ENTITIES == Integer.MAX_VALUE
-                && MAX_HISTORY == Long.MAX_VALUE
+                && MAX_HISTORY == Integer.MAX_VALUE
                 && INVENTORY_MODE == 0
                 && SPEED_REDUCTION == 0
                 && FAST_PLACEMENT
